Rename package for neoforge
