+++ NeoForge/src/main/java/dev/lukebemish/defaultresources/impl/neoforge/package-info.java
@@ -1,6 +1,6 @@
 @NullMarked
 @ApiStatus.Internal
-package dev.lukebemish.defaultresources.impl.forge;
+package dev.lukebemish.defaultresources.impl.neoforge;
 
 import org.jetbrains.annotations.ApiStatus;
 import org.jspecify.annotations.NullMarked;
